fix config syntax

+++ config/unicorn.rb
@@ -23,13 +23,7 @@ GC.respond_to?(:copy_on_write_friendly=) and GC.copy_on_write_friendly = true
 before_fork do |server, worker|
   defined?(ActiveRecord::Base) and
       ActiveRecord::Base.connection.disconnect!
-
-  if File.exists?(old_pid) && server.pid != old_pid
-    begin
-      Process.kill("QUIT", File.read(old_pid).to_i)
-    rescue Errno::ENOENT, Errno::ESRCH
-    end
-  end
+      Process.kill("QUIT", File.read(old_pid).to_i) if File.exists?(old_pid) && server.pid != old_pid
 end
 
 after_fork do |server, worker|
